updated Rating type from Integer to Float + added rating query

+++ graphql/types/Rating/types.ts
@@ -4,7 +4,7 @@ export const CreateUserRatingsInput = inputObjectType({
   name: "CreateUserRatingsInput",
   definition: (t) => {
     t.nonNull.string("playerId");
-    t.nonNull.int("rating");
+    t.nonNull.float("rating");
   },
 });
